scala ast: add typed nodes
