+++ tools/json-ast/x/scala/inspect.go
@@ -8,13 +8,6 @@ import (
 	tscala "github.com/smacker/go-tree-sitter/scala"
 )
 
-// Node represents a tree-sitter node.
-type Node struct {
-	Type     string `json:"type"`
-	Text     string `json:"text,omitempty"`
-	Children []Node `json:"children,omitempty"`
-}
-
 // Program represents a parsed Scala source file.
 type Program struct {
 	Root Node `json:"root"`
@@ -28,22 +21,6 @@ func Inspect(src string) (*Program, error) {
 	if err != nil {
 		return nil, fmt.Errorf("parse: %w", err)
 	}
-	root := convertNode(tree.RootNode(), []byte(src))
+	root := convert(tree.RootNode(), []byte(src))
 	return &Program{Root: root}, nil
-}
-
-func convertNode(n *sitter.Node, src []byte) Node {
-	node := Node{Type: n.Type()}
-	if n.ChildCount() == 0 {
-		node.Text = n.Content(src)
-		return node
-	}
-	for i := 0; i < int(n.ChildCount()); i++ {
-		child := n.Child(i)
-		if child == nil {
-			continue
-		}
-		node.Children = append(node.Children, convertNode(child, src))
-	}
-	return node
 }
